+++ src/main/java/com/backend/domain/route/service/RouteService.java
@@ -9,7 +9,7 @@ import com.backend.domain.route.entity.Route;
 import com.backend.domain.route.repository.RouteRepository;
 import com.backend.domain.routePlace.entity.RoutePlace;
 import com.backend.domain.routePlace.repository.RoutePlaceRepository;
-import com.backend.global.security.util.MemberUtil;
+import com.backend.global.util.MemberUtil;
 import com.backend.global.util.FormatUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
